Move the span format into its own module
The wrapper carried the answer grammar inline: fourteen state constants, a
transition function, the constrained decoder and the parser, none of which
concern prompting or calling a model. prefix_allowed_tokens_fn now meets a
named class across a module boundary.

Grammar and parser move together because they describe one format in two
directions. A change to what training emits lands in one file, and drift
between what we constrain and what we read back stays visible.

Pure refactor: the same 68 tests pass, split to mirror the modules.
